pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "groq>=0.30.0",
    "openai>=1.99.1",
    "python-dotenv>=1.1.1",
]
